Kompletni refaktor struktury repozitare.
 # Co:
1. Rozdeleni logiky z jednoho souboru na vice souboru.
2. Zmena struktury repozitare (Rozdeleni do slozek.).
 # Proc:
1. Atomictejsi rozdeleni. Jedna myslena jeden souboru.
2. Vice slozek pro specificke urceni. Oddeleni skupin souboru.
